+++ src/esperanza/checkpoint.h
@@ -27,7 +27,7 @@ class Checkpoint {
   std::map<uint32_t, uint64_t> m_prevDynastyVotes;
 
   // Set of validatorAddresses for validators that voted that checkpoint
-  std::set<uint256> m_voteSet;
+  std::set<uint160> m_voteSet;
 
   uint64_t GetCurDynastyVotes(uint32_t epoch);
   uint64_t GetPrevDynastyVotes(uint32_t epoch);
