+++ src/esperanza/validation.h
@@ -40,7 +40,7 @@ bool CheckAdminTransaction(CValidationState &state, const CTransaction &tx,
 //! \param tx
 //! \param validatorAddressOut
 //! \return true if successful, false otherwise.
-bool ExtractValidatorIndex(const CTransaction &tx, uint256 &validatorAddressOut);
+bool ExtractValidatorIndex(const CTransaction &tx, uint160 &validatorAddressOut);
 
 }  // namespace esperanza
